Network.TypedProtocol.Codec

Defining and using Codecs

Codec type

data Codec ps failure m bytes

A codec for a Protocol handles the encoding and decoding of typed protocol messages. This is typically used when sending protocol messages over untyped channels. The codec chooses the exact encoding, for example encoding in some text-based syntax, or some choice of binary format.

The codec is parametrised by:

The protocol

the type of decoding failures

the monad in which the decoder runs

the type of the encoded data (typically strings or bytes)

A codec consists of a message encoder and a decoder.

The encoder is supplied both with the message to encode and the current protocol state (matching the message). The protocol state can be either a client or server state, but for either peer role it is a protocol state in which the peer has agency, since those are the only states where a peer needs to encode a message to be able to send it.

For example a simple text encoder for the ping/pong protocol could be:

encode :: SingI st
       => Message PingPong st st'
       -> String
encode MsgPing = "ping\n"
encode MsgDone = "done\n"
encode MsgPong = "pong\n"

The decoder is also given the current protocol state and it is expected to be able to decode any message that is valid in that state, but only messages that are valid in that state. Messages that are unexpected for the current state should be treated like any other decoding format error.

While the current protocol state is known, the state that the message will have the peer transition to is not known. For this reason the decoded message is wrapped in the SomeMessage constructor which hides the "to" state.

The decoder uses an incremental decoding interface DecodeStep so that input can be supplied (e.g. from a Channel) bit by bit. This style of decoder allows but does not require a format with message framing where the decoder input matches exactly with the message boundaries.

decode :: forall st m. SingI st
       => StateToken st
       -> m (DecodeStep String String m (SomeMessage st))
decode stok =
  decodeTerminatedFrame '\n' $ \str trailing ->
    case (stok, str) of
      (SingBusy, "pong") ->
           DecodeDone (SomeMessage MsgPong) trailing
      (SingIdle, "ping") ->
           DecodeDone (SomeMessage MsgPing) trailing
      (SingIdle, "done") ->
           DecodeDone (SomeMessage MsgDone) trailing
      _ -> DecodeFail ("unexpected message: " ++ str)

See "typed-protocols-examples" for the full example.

Note that the pattern matching on the combination of the message string and the protocol state. This neatly fulfils the requirement that we only return messages that are of the correct type for the given protocol state.

This toy example format uses newlines n as a framing format. See DecodeStep for suggestions on how to use it for more realistic formats.

Constructors

Codec

Fields

encode :: forall (st :: ps) (st' :: ps). StateTokenI st => ActiveState st => Message ps st st' -> bytes

decode :: forall (st :: ps). ActiveState st => StateToken st -> m (DecodeStep bytes failure m (SomeMessage st))

hoistCodec

Arguments

:: Functor n

=> (forall x. m x -> n x)
a natural transformation

-> Codec ps failure m bytes

-> Codec ps failure n bytes

Change functor in which the codec is running.

isoCodec

Arguments

:: Functor m

=> (bytes -> bytes')
map from bytes to bytes'

-> (bytes' -> bytes)
its inverse

-> Codec ps failure m bytes
codec

-> Codec ps failure m bytes'

Change bytes of a codec.

mapFailureCodec

Arguments

:: Functor m

=> (failure -> failure')
a function to apply to failure

-> Codec ps failure m bytes

-> Codec ps failure' m bytes

Modify failure type.

Incremental decoding

data DecodeStep bytes failure m a

An incremental decoder with return a value of type a.

This interface is not designed to be used directly for implementing decoders, only for running them. In real applications it is expected to use libraries for text or binary decoding and to implement appropriate wrappers to match up with this incremental decoder interface.

This style of interface already closely matches that provided by libraries such as attoparsec for text formats, and binary, cereal and cborg for binary formats.

Constructors

DecodePartial (Maybe bytes -> m (DecodeStep bytes failure m a))
The decoder has consumed the available input and needs more to continue. Provide Just if more input is available and Nothing otherwise, and you will get a new DecodeStep.

DecodeDone a (Maybe bytes)
The decoder has successfully finished. This provides the decoded result value plus any unused input.

DecodeFail failure
The decoder ran into an error. The decoder either used fail or was not provided enough input.

runDecoder

Arguments

:: Monad m

=> [bytes]
bytes to be fed into the incremental DecodeStep

-> DecodeStep bytes failure m a
decoder

-> m (Either failure a)

Run a codec incremental decoder DecodeStep against a list of input.

It ignores any unused trailing data. This is useful for demos, quick experiments and tests.

See also runDecoderWithChannel

runDecoderPure

Arguments

:: Monad m

=> (forall b. m b -> b)
run monad m in a pure way, e.g. runIdentity

-> m (DecodeStep bytes failure m a)

-> [bytes]
input bytes

-> Either failure a

A variant of runDecoder that is suitable for "pure" monads that have a run function. This includes ST, using runST.

hoistDecodeStep

Arguments

:: Functor n

=> (forall x. m x -> n x)
a natural transformation

-> DecodeStep bytes failure m a

-> DecodeStep bytes failure n a

Change functor in which the codec is running.

isoDecodeStep

Arguments

:: Functor m

=> (bytes -> bytes')
map from bytes to bytes'

-> (bytes' -> bytes)

-> DecodeStep bytes failure m a

-> DecodeStep bytes' failure m a

Change bytes of DecodeStep.

mapFailureDecodeStep

Arguments

:: Functor m

=> (failure -> failure')
a function to apply to failure

-> DecodeStep bytes failure m a

-> DecodeStep bytes failure' m a

Modify failure type.

Related types

SomeMessage

data SomeMessage (st :: ps) where

When decoding a Message we only know the expected "from" state. We cannot know the "to" state as this depends on the message we decode. To resolve this we use the SomeMessage wrapper which uses an existential type to hide the "to" state.

Constructors

SomeMessage :: (StateTokenI st, StateTokenI st', ActiveState st) => Message ps st st' -> SomeMessage st

StateToken

type family StateToken :: ps -> Type

A type family for protocol state token, e.g. term level representation of type level state (also known as singleton).

This type family is similar to Sing type class in the "singletons" package, but specific for protocol states.

class StateTokenI st where

A type class which hides a state token / singleton inside a class dictionary.

This is similar to the SingI instance, but specific to protocol state singletons.

Methods

stateToken :: StateToken st

ActiveState

class IsActiveState st (agency :: Agency) where

A type class which restricts states to ones that have ClientAgency or ServerAgency, excluding NobodyAgency.

One can use notActive to eliminate cases for which both IsActiveState st is in scope and for which we have an evidence that the state is not active (i.e. a singleton). This is useful when writing a Codec.

Methods

activeAgency :: ActiveAgency' st agency

type ActiveState st = IsActiveState st (StateAgency st)

A constraint which provides an evidence that the protocol isn't in a terminal state.

type ActiveAgency st = ActiveAgency' st (StateAgency st)

Evidence that the protocol isn't in a terminal state.

data ActiveAgency' st agency where

Evidence that one side of the protocol has the agency, and thus that the protocol hasn't yet terminated.

Constructors

ClientHasAgency :: StateAgency st ~ ClientAgency => ActiveAgency' st ClientAgency
Evidence that the client has the agency.

ServerHasAgency :: StateAgency st ~ ServerAgency => ActiveAgency' st ServerAgency
Evidence that the server has the agency.

notActiveState :: forall ps (st :: ps). StateAgency st ~ NobodyAgency => ActiveState st => StateToken st -> forall a. a

This is useful function to eliminate cases where the `ActiveState st` is provided but we are given a state in which neither side has agency (NobodyAgency). For example when writing a codec, we only need to encode / decode messages which are in active states, but to make such functions total, notActiveState needs to be used to eliminate the states in which nobody has agency.

A good analogy for this function is absurd :: Void -> a.

PeerRole

data PeerRole

Types for client and server peer roles. As protocol can be viewed from either client or server side.

Note that technically "client" and "server" are arbitrary labels. The framework is completely symmetric between the two peers.

This definition is only used as promoted types and kinds, never as values.

Constructors

AsClient

AsServer

CodecFailure

data CodecFailure

Each Codec can use whatever failure type is appropriate. This simple exception type is provided for use by simple codecs (e.g. "identity") when nothing more than a String is needed. It is an instance of Exception.

Constructors

CodecFailureOutOfInput

CodecFailure String

Testing codec properties

data AnyMessage ps where

Any message for a protocol, with a StateTokenI constraint which gives access to protocol state.

Used where we don't know statically what the state type is, but need the agency and message to match each other.

Constructors

AnyMessage

Fields

:: forall ps (st :: ps) (st' :: ps). (StateTokenI st, ActiveState st)

=> Message ps (st :: ps) (st' :: ps)

Message between some states

-> AnyMessage ps

Bundled Patterns

pattern AnyMessageAndAgency :: forall ps. () => forall (st :: ps) (st' :: ps). (StateTokenI st, ActiveState st) => StateToken st -> Message ps st st' -> AnyMessage ps
A convenient pattern synonym which unwrap AnyMessage giving both the singleton for the state and the message.

prop_codecM

Arguments

:: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps))

=> Codec ps failure m bytes
codec

-> AnyMessage ps
some message

-> m Bool
returns True iff round trip returns the exact same message

The Codec round-trip property: decode after encode gives the same message. Every codec must satisfy this property.

prop_codec :: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps)) => (forall a. m a -> a) -> Codec ps failure m bytes -> AnyMessage ps -> Bool

The Codec round-trip property in a pure monad.

prop_codec_splitsM

Arguments

:: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps))

=> (bytes -> [[bytes]])
alternative re-chunkings of serialised form

-> Codec ps failure m bytes

-> AnyMessage ps

-> m Bool

A variant on the codec round-trip property: given the encoding of a message, check that decode always gives the same result irrespective of how the chunks of input are fed to the incremental decoder.

This property guards against boundary errors in incremental decoders. It is not necessary to check this for every message type, just for each generic codec construction. For example given some binary serialisation library one would write a generic adaptor to the codec interface. This adaptor has to deal with the incremental decoding and this is what needs to be checked.

prop_codec_splits

Arguments

:: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps))

=> (bytes -> [[bytes]])
alternative re-chunkings of serialised form

-> (forall a. m a -> a)

-> Codec ps failure m bytes

-> AnyMessage ps

-> Bool

Like prop_codec_splitsM but run in a pure monad m, e.g. Identity.

prop_codec_binary_compatM

Arguments

:: forall psA psB failure m bytes. (Monad m, Eq (AnyMessage psA))

=> Codec psA failure m bytes

-> Codec psB failure m bytes

-> (forall (stA :: psA). ActiveState stA => StateToken stA -> SomeState psB)
the states of A map directly to states of B.

-> AnyMessage psA

-> m Bool

Binary compatibility of two protocols

We check the following property:

Using codec A, we encode a message of protocol psA to bytes.

When we decode those bytes using codec B, we get a message of protocol psB.

When we encode that message again using codec B, we get bytes.

When we decode those bytes using codec A, we get the original message again.

prop_codec_binary_compat

Arguments

:: forall psA psB failure m bytes. (Monad m, Eq (AnyMessage psA))

=> (forall a. m a -> a)

-> Codec psA failure m bytes

-> Codec psB failure m bytes

-> (forall (stA :: psA). StateToken stA -> SomeState psB)
the states of A map directly to states of B.

-> AnyMessage psA

-> Bool

Like prop_codec_splitsM but run in a pure monad m, e.g. Identity.

prop_codecs_compatM

Arguments

:: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps), forall a. Monoid a => Monoid (m a))

=> Codec ps failure m bytes
first codec

-> Codec ps failure m bytes
second codec

-> AnyMessage ps
some message

-> m Bool

Compatibility between two codecs of the same protocol. Encode a message with one codec and decode it with the other one, then compare if the result is the same as initial message.

prop_codecs_compat :: forall ps failure m bytes. (Monad m, Eq (AnyMessage ps), forall a. Monoid a => Monoid (m a)) => (forall a. m a -> a) -> Codec ps failure m bytes -> Codec ps failure m bytes -> AnyMessage ps -> Bool

Like prop_codecs_compatM but run in a pure monad m, e.g. Identity.

data SomeState (ps :: Type) where

Auxiliary definition for prop_codec_binary_compatM.

Used for the existential st :: ps parameter when expressing that for each value of PeerHasAgency for protocol A, there is a corresponding PeerHasAgency for protocol B of some st :: ps.

Constructors

SomeState

Fields

:: forall ps (st :: ps). ActiveState st

=> StateToken st

state token for some active state st

-> SomeState ps