Network.TypedProtocol.Peer.Server

Description

Bidirectional patterns for Peer ps AsServer. The advantage of these patterns is that they automatically provide the ReflRelativeAgency evidence.

Server type alias and its pattern synonyms

type Server ps pl st m a = Peer ps AsServer pl st m a

pattern Effect

Arguments

:: forall ps pl st m a. m (Server ps pl st m a)
monadic continuation

-> Server ps pl st m a

Server role pattern for Effect.

pattern Yield

Arguments

:: forall ps pl st m a. forall st'. (StateTokenI st, StateTokenI st', StateAgency st ~ ServerAgency, Outstanding pl ~ Z)

=> Message ps st st'
protocol message

-> Server ps pl st' m a
continuation

-> Server ps pl st m a

Server role pattern for Yield

pattern Await

Arguments

:: forall ps pl st m a. (StateTokenI st, StateAgency st ~ ClientAgency, Outstanding pl ~ Z)

=> (forall st'. Message ps st st' -> Server ps pl st' m a)
continuation

-> Server ps pl st m a

Server role pattern for Await

pattern Done

Arguments

:: forall ps pl st m a. (StateTokenI st, StateAgency st ~ NobodyAgency, Outstanding pl ~ Z)

=> a
protocol return value

-> Server ps pl st m a

Server role pattern for Done

pattern YieldPipelined

Arguments

:: forall ps st n c m a. forall st' st''. (StateTokenI st, StateTokenI st', StateAgency st ~ ServerAgency)

=> Message ps st st'
pipelined message

-> Receiver ps st' st'' m c

-> Server ps (Pipelined (S n) c) st'' m a
continuation

-> Server ps (Pipelined n c) st m a

Server role pattern for YieldPipelined

pattern Collect

Arguments

:: forall ps st n c m a. (StateTokenI st, ActiveState st)

=> Maybe (Server ps (Pipelined (S n) c) st m a)
continuation, executed if no message has arrived so far

-> (c -> Server ps (Pipelined n c) st m a)
continuation

-> Server ps (Pipelined (S n) c) st m a

Server role pattern for Collect

Receiver type alias and its pattern synonyms

type Receiver ps st stdone m c = Receiver ps AsServer st stdone m c

pattern ReceiverEffect :: forall ps st stdone m c. m (Receiver ps st stdone m c) -> Receiver ps st stdone m c

pattern ReceiverAwait :: forall ps st stdone m c. () => (StateTokenI st, ActiveState st, StateAgency st ~ ClientAgency) => (forall st'. Message ps st st' -> Receiver ps st' stdone m c) -> Receiver ps st stdone m c

pattern ReceiverDone :: forall ps stdone m c. c -> Receiver ps stdone stdone m c

ServerPipelined type alias and its pattern synonym

type ServerPipelined ps st m a = PeerPipelined ps AsServer st m a

A description of a peer that engages in a protocol in a pipelined fashion.

data PeerPipelined ps pr (st :: ps) m a where

A description of a peer that engages in a protocol in a pipelined fashion.

This type is useful for wrapping pipelined peers to hide information which is only relevant in peer lift. It is expected by runPeerPipelinedWithDriver.

Bundled Patterns

pattern ServerPipelined :: forall ps st m a. forall c. Server ps (Pipelined Z c) st m a -> ServerPipelined ps st m a

re-exports

data IsPipelined where

Promoted data type which indicates if Peer is used in pipelined mode or not.

Constructors

Pipelined :: N -> Type -> IsPipelined
Pipelined peer which is using `c :: Type` for collecting responses from a pipelined messages. N indicates depth of pipelining.

NonPipelined :: IsPipelined
Non-pipelined peer.

type family Outstanding pl where ...

Type level count of the number of outstanding pipelined yields for which we have not yet collected a receiver result. Used to ensure that Collect is only used when there are outstanding results to collect (e.g. after YieldPipeliend was used); and to ensure that the non-pipelined primitives Yield, Await and Done are only used when there are none unsatisfied pipelined requests.

Equations

Outstanding 'NonPipelined = Z

Outstanding ('Pipelined n _) = n

data N

A type level inductive natural number.

Constructors

Z

S N

data Nat (n :: N) where

A value level inductive natural number, indexed by the corresponding type level natural number N.

This is often needed when writing pipelined peers to be able to count the number of outstanding pipelined yields, and show to the type checker that Collect and Done are being used correctly.

Bundled Patterns

pattern Succ :: () => m ~ S n => Nat n -> Nat m

pattern Zero :: () => Z ~ n => Nat n