Network.TypedProtocol.Proofs

Description

Proofs and helpful testing utilities.

Connect proofs

connect

Arguments

:: forall ps (pr :: PeerRole) (initSt :: ps) m a b. (Monad m, SingI pr)

=> Peer ps pr NonPipelined initSt m a
a peer

-> Peer ps (FlipAgency pr) NonPipelined initSt m b
a peer with flipped agency

-> m (a, b, TerminalStates ps)
peers results and an evidence of their termination

The connect function takes two peers that agree on a protocol and runs them in lock step, until (and if) they complete.

The connect function serves a few purposes.

The fact we can define this function at at all proves some minimal sanity property of the typed protocol framework.

It demonstrates that all protocols defined in the framework can be run with synchronous communication rather than requiring buffered communication.

It is useful for testing peer implementations against each other in a minimalistic setting.

connectPipelined

Arguments

:: forall ps (pr :: PeerRole) (st :: ps) m a b. (Monad m, SingI pr)

=> [Bool]
an interleaving

-> PeerPipelined ps pr st m a
a pipelined peer

-> Peer ps (FlipAgency pr) NonPipelined st m b
a non-pipelined peer with fliped agency

-> m (a, b, TerminalStates ps)
peers results and an evidence of their termination

Analogous to connect but also for pipelined peers.

Since pipelining allows multiple possible interleavings, we provide a [Bool] parameter to control the choices. Each True will trigger picking the first choice in the SenderCollect construct (if possible), leading to more results outstanding. This can also be interpreted as a greater pipeline depth, or more messages in-flight.

This can be exercised using a QuickCheck style generator.

data TerminalStates ps where

The terminal states for the protocol. Used in connect and connectPipelined to return the states in which the peers terminated.

Constructors

TerminalStates

Fields

:: forall ps (st :: ps). StateAgency st ~ NobodyAgency

=> StateToken st

state termination evidence for the first peer

-> StateToken st

state termination evidence for the second peer

-> TerminalStates ps

Pipelining proofs

Additional proofs specific to the pipelining features

forgetPipelined

Arguments

:: forall ps (pr :: PeerRole) (st :: ps) m a. Functor m

=> [Bool]
interleaving choices for pipelining allowed by Collect primitive. False values or `[]` give no pipelining.

-> PeerPipelined ps pr st m a

-> Peer ps pr NonPipelined st m a

Proof that we have a total conversion from pipelined peers to regular peers. This is a sanity property that shows that pipelining did not give us extra expressiveness or to break the protocol state machine.

promoteToPipelined

Arguments

:: forall ps (pr :: PeerRole) st m a. Functor m

=> Peer ps pr NonPipelined st m a
a peer

-> PeerPipelined ps pr st m a
a pipelined peer

Promote a peer to a pipelined one.

This is a right inverse of forgetPipelined, e.g.

>>> forgetPipelined . promoteToPipelined = id

Pipeline proof helpers

data Queue (n :: N) a where

A size indexed queue. This is useful for proofs, including connectPipelined but also as so-called direct functions for running a client and server wrapper directly against each other.

Constructors

EmptyQ :: Queue Z a

ConsQ :: a -> Queue n a -> Queue (S n) a

enqueue :: a -> Queue n a -> Queue (S n) a

At an element to the end of a Queue. This is not intended to be efficient. It is only for proofs and tests.

Auxiliary functions

pipelineInterleaving

Arguments

:: Int
Bound on outstanding responses

-> [Bool]
Pipelining choices

-> [req]

-> [resp]

-> [Either req resp]

A reference specification for interleaving of requests and responses with pipelining, where the environment can choose whether a response is available yet.

This also supports bounded choice where the maximum number of outstanding in-flight responses is limited.